Created a profile directory  and moved all the profile components in it.
