+++ src/components/pages/profile/InfoList.js
@@ -1,5 +1,5 @@
 import React from "react";
-import profileInfoStyle from "./profileStyles/profileInfoStyle";
+import profileInfoStyle from "../profileStyles/profileInfoStyle";
 import List from "@material-ui/core/List";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemText from "@material-ui/core/ListItemText";
