+++ src/components/pages/profile/Profile.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useContext } from "react";
-import NavBar from "../Navbar.js";
+import NavBar from "../../Navbar.js";
 import ProfileInfo from "./ProfileInfo";
 import AddUser from "./AddProfile.js";
 //import EditUser from "./EditUser";
@@ -7,8 +7,8 @@ import AddUser from "./AddProfile.js";
 import Container from "@material-ui/core/Container";
 import Grid from "@material-ui/core/Grid";
 
-import ProfileContext from "../../contexts/ProfileContext";
-import { createProfileContext } from "../../functions";
+import ProfileContext from "../../../contexts/ProfileContext";
+import { createProfileContext } from "../../../functions";
 
 export default function Profile(props) {
