+++ src/components/pages/profile/SelectState.js
@@ -1,10 +1,10 @@
 import React from "react";
-import formStyles from "./profileStyles/formStyles";
+import formStyles from "../profileStyles/formStyles";
 import InputLabel from "@material-ui/core/InputLabel";
 import MenuItem from "@material-ui/core/MenuItem";
 import FormControl from "@material-ui/core/FormControl";
 import Select from "@material-ui/core/Select";
-import states from './data/states';
+import states from '../data/states';
 
 function SelectState({addHandleChange}) {
     const classes = formStyles();
